+++ apps/site/components/EOL/EOLReleaseTable/index.tsx
@@ -1,10 +1,11 @@
+import { getTranslations } from 'next-intl/server';
 import type { FC } from 'react';
 
 import provideReleaseData from '#site/next-data/providers/releaseData';
 import provideVulnerabilities from '#site/next-data/providers/vulnerabilities';
 import { EOL_VERSION_IDENTIFIER } from '#site/next.constants.mjs';
 
-import EOLReleaseTableClient from './TableClient';
+import EOLReleaseTableBody from './TableBody';
 
 const EOLReleaseTable: FC = async () => {
   const releaseData = await provideReleaseData();
@@ -14,11 +15,27 @@ const EOLReleaseTable: FC = async () => {
     release => release.status === EOL_VERSION_IDENTIFIER
   );
 
+  const t = await getTranslations();
+
   return (
-    <EOLReleaseTableClient
-      eolReleases={eolReleases}
-      vulnerabilities={vulnerabilities}
-    />
+    <table id="tbVulnerabilities">
+      <thead>
+        <tr>
+          <th>
+            {t('components.eolTable.version')} (
+            {t('components.eolTable.codename')})
+          </th>
+          <th>{t('components.eolTable.lastUpdated')}</th>
+          <th>{t('components.eolTable.vulnerabilities')}</th>
+          <th>{t('components.eolTable.details')}</th>
+        </tr>
+      </thead>
+
+      <EOLReleaseTableBody
+        eolReleases={eolReleases}
+        vulnerabilities={vulnerabilities}
+      />
+    </table>
   );
 };
